+++ website-nextjs/src/components/admin/benchmark-detail/BenchmarkTableResult.tsx
@@ -198,70 +198,6 @@ const BenchmarkTableResult = () => {
           </tbody>
         </table>
       </div>
-      {/* Pagination */}
-      <div className="flex text-xs items-center gap-2 mt-4 justify-between">
-        <div className="text-dark-grey">
-          Showing{" "}
-          <span className="font-bold">
-            {" "}
-            {currentPage === totalPages
-              ? memoizedMetaData.length
-              : table.getState().pagination.pageSize * currentPage}
-          </span>{" "}
-          of <span className="font-bold"> {memoizedMetaData.length} </span>
-        </div>
-        <div>
-          <div className="flex items-center gap-2">
-            <button
-              className={`flex gap-2 items-center rounded p-1 ${
-                !table.getCanPreviousPage() ? " text-dark-grey" : " text-navy"
-              }`}
-              onClick={() => table.previousPage()}
-              disabled={!table.getCanPreviousPage()}
-            >
-              <ArrowRightIcon
-                className={`h-3 w-3 rotate-180 ${
-                  !table.getCanPreviousPage() ? "stroke-dark-grey" : ""
-                }`}
-                fill="none"
-              />
-              Previous
-            </button>
-            {Array.from({ length: table.getPageCount() }, (_, i) => i + 1).map(
-              (page) => (
-                <button
-                  key={page}
-                  className={`border border-stroke rounded py-1 px-2.5 ${
-                    page - 1 === table.getState().pagination.pageIndex
-                      ? "bg-navy text-white"
-                      : "text-dark-green"
-                  }`}
-                  onClick={() => table.setPageIndex(page - 1)}
-                  disabled={page - 1 === table.getState().pagination.pageIndex}
-                >
-                  {page}
-                </button>
-              )
-            )}
-            <button
-              className={`flex gap-2 items-center rounded p-1
-                ${!table.getCanNextPage() ? " text-dark-grey" : " text-navy"}
-                `}
-              onClick={() => table.nextPage()}
-              disabled={!table.getCanNextPage()}
-            >
-              Next
-              <ArrowRightIcon
-                className={`h-3 w-3 rotate-145 ${
-                  !table.getCanNextPage() ? "stroke-dark-grey" : "stroke-navy"
-                }`}
-                fill="none"
-              />
-            </button>
-          </div>
-        </div>
-        <div></div>
-      </div>
     </div>
   )
 }
