+++ website-nextjs/src/components/shared/tables/FilterTable.tsx
@@ -1,0 +1,86 @@
+import DebouncedInput from "@/components/admin/raw-result/DebouncedInput"
+import FilterAutoComplete from "@/components/admin/raw-result/FilterAutoComplete"
+import { BenchmarkResult } from "@/types/benchmark"
+import { Column } from "@tanstack/react-table"
+import { useMemo } from "react"
+
+export default function Filter({
+  column,
+}: {
+  column: Column<BenchmarkResult, unknown>
+}) {
+  const { filterVariant } =
+    (column.columnDef.meta as { filterVariant?: string }) ?? {}
+
+  const columnFilterValue = column.getFilterValue()
+
+  const sortedUniqueValues = useMemo(
+    () =>
+      filterVariant === "range"
+        ? []
+        : Array.from(column.getFacetedUniqueValues().keys())
+            .sort()
+            .slice(0, 5000),
+    [column.getFacetedUniqueValues(), filterVariant]
+  )
+
+  return filterVariant === "range" ? (
+    <div>
+      <div className="flex space-x-2">
+        <DebouncedInput
+          type="number"
+          min={Number(column.getFacetedMinMaxValues()?.[0] ?? "")}
+          max={Number(column.getFacetedMinMaxValues()?.[1] ?? "")}
+          value={(columnFilterValue as [number, number])?.[0] ?? ""}
+          onChange={(value) =>
+            column.setFilterValue((old: [number, number]) => [value, old?.[1]])
+          }
+          placeholder={`Min ${
+            column.getFacetedMinMaxValues()?.[0] !== undefined
+              ? `(${column.getFacetedMinMaxValues()?.[0]})`
+              : ""
+          }`}
+          className="w-24 border rounded px-2 h-8"
+        />
+        <DebouncedInput
+          type="number"
+          min={Number(column.getFacetedMinMaxValues()?.[0] ?? "")}
+          max={Number(column.getFacetedMinMaxValues()?.[1] ?? "")}
+          value={(columnFilterValue as [number, number])?.[1] ?? ""}
+          onChange={(value) =>
+            column.setFilterValue((old: [number, number]) => [old?.[0], value])
+          }
+          placeholder={`Max ${
+            column.getFacetedMinMaxValues()?.[1]
+              ? `(${column.getFacetedMinMaxValues()?.[1]})`
+              : ""
+          }`}
+          className="w-24 border rounded px-2 h-8"
+        />
+      </div>
+      <div className="h-1" />
+    </div>
+  ) : filterVariant === "select" ? (
+    <select
+      onChange={(e) => column.setFilterValue(e.target.value)}
+      value={columnFilterValue?.toString()}
+    >
+      <option value="">All</option>
+      {sortedUniqueValues.map((value) => (
+        <option value={value} key={value}>
+          {value}
+        </option>
+      ))}
+    </select>
+  ) : (
+    <>
+      <FilterAutoComplete
+        options={sortedUniqueValues.map((val) => ({
+          value: val,
+          label: val,
+        }))}
+        setFilterValue={column.setFilterValue}
+      />
+    </>
+  )
+}
