+++ website-nextjs/src/pages/dashboard/raw-result.tsx
@@ -17,7 +17,11 @@ const PagePerformanceHistory = () => {
       </Head>
       <div className="bg-light-blue h-screen">
         <Navbar />
-        <div className={`px-6 ${isNavExpanded ? "ml-64" : "ml-20"}`}>
+        <div
+          className={`px-6 min-h-[calc(100vh-120px)] ${
+            isNavExpanded ? "ml-64" : "ml-20"
+          }`}
+        >
           <AdminHeader />
           {/* Content */}
           <DetailSection />
